Work on parsing and optimization
We are being a bit consistent with what things parse to what.
In particular, we have flat blocks (blocks that don't call functions)
having this distinction is very useful for optimization. This allows
us to take the mandelbrot benchmark from 15s to 3.5s on my machine. (
compiled using gcc -O0)